fix: refactor python wrappers to use common class
